pyproject.toml
[tool.poetry]
name = "uglygpt"
version = "0.1.0"
description = ""
authors = ["uglyboy-tl <25730357+uglyboy-tl@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uglychain = {extras = ["llama-index","bm25"], version = "^0.1.0"}
llama-index-readers-github = {version = "^0.1.1", python = ">=3.11,<3.12", extras = ["github"], optional = true}

[tool.poetry.extras]
github = ["llama-index-readers-github"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"